feat(core): add optimization into findOneInChannel and more refactoring
